requirements.txt
numpy>=1.24
Pillow>=9.5
scikit-image>=0.21
scikit-learn>=1.3
PyEMD>=0.5.1
matplotlib>=3.7
scipy>=1.10
EMD-signal>=0.6.6
rich>=13.7
torch>=2.2
torchvision>=0.17

